Fix error when change card between status columns
This error consists in when a task card is moved to one status that doesn't exists on board, the backend returns one error response but the frontend shows the card changing normally. This make user think that the card has been successfully moved, but that is not the case when the user refreshes the page.